feat: support flystem v2 and guzzle 6.3
added support for flysystem v2 and guzzle 6.3 and automatic handling the providers to support the versions
